+++ api/test/h3-data/mocks/h3-risk-map.mock.ts
@@ -1,24 +1,17 @@
-import { Unit } from '../../../src/modules/units/unit.entity';
-import { UnitConversion } from '../../../src/modules/unit-conversions/unit-conversion.entity';
-import { Indicator } from '../../../src/modules/indicators/indicator.entity';
+import { Unit } from 'modules/units/unit.entity';
+import { UnitConversion } from 'modules/unit-conversions/unit-conversion.entity';
+import { Indicator } from 'modules/indicators/indicator.entity';
 import { h3DataMock } from './h3-data.mock';
-import { Material } from '../../../src/modules/materials/material.entity';
-import {
-  createGeoRegion,
-  createIndicatorRecord,
-  createMaterial,
-  createSourcingLocation,
-  createSourcingRecord,
-} from '../../entity-mocks';
-import { GeoRegion } from '../../../src/modules/geo-regions/geo-region.entity';
-import { SourcingLocation } from '../../../src/modules/sourcing-locations/sourcing-location.entity';
-import { SourcingRecord } from '../../../src/modules/sourcing-records/sourcing-record.entity';
-import { H3Data } from '../../../src/modules/h3-data/h3-data.entity';
+import { Material } from 'modules/materials/material.entity';
+import { createMaterial, createMaterialToH3 } from '../../entity-mocks';
+import { H3Data } from 'modules/h3-data/h3-data.entity';
+import { MATERIAL_TO_H3_TYPE } from 'modules/materials/material-to-h3.entity';
 
 export const createWorldForRiskMapGeneration = async (data: {
   indicatorType: string;
   fakeTable: string;
   fakeColumn: string;
+  year: number;
 }): Promise<{
   indicator: Indicator;
   material: Material;
@@ -42,17 +35,21 @@ export const createWorldForRiskMapGeneration = async (data: {
   indicator.nameCode = data.indicatorType;
   await indicator.save();
 
-  const h3Data = await h3DataMock(
-    data.fakeTable,
-    data.fakeColumn,
-    null,
-    indicator.id,
-  );
+  const h3Data = await h3DataMock({
+    h3TableName: data.fakeTable,
+    h3ColumnName: data.fakeColumn,
+    indicatorId: indicator.id,
+    year: data.year,
+  });
 
   const material = await createMaterial({
     name: 'Material with no H3',
-    harvestId: h3Data.id,
-    producerId: h3Data.id,
   });
+  await createMaterialToH3(
+    material.id,
+    h3Data.id,
+    MATERIAL_TO_H3_TYPE.PRODUCER,
+  );
+  await createMaterialToH3(material.id, h3Data.id, MATERIAL_TO_H3_TYPE.HARVEST);
   return { indicator, material, h3Data, unit, unitConversion };
 };